+++ packages/core/ui-svelte/src/lib/ui/shadcn-workaround/select.workaround.svelte
@@ -9,14 +9,16 @@ import { cn } from '$lib/utils/shadcn.js'
 
 let {
 	value = $bindable(),
+	disabled = false,
 	class: className,
 	options,
 	placeholder,
 	handleChange
 }: {
-	value: string | undefined
+	value: string | number | undefined
+	disabled?: boolean
 	class?: string
-	options?: { value: string; label: string }[]
+	options?: { value: string | number; label: string | number }[]
 	placeholder?: string
 	handleChange?: (event: Event) => void
 } = $props()
@@ -25,6 +27,7 @@ let {
 <select
 bind:value={value}
 onchange={handleChange}
+disabled={disabled}
 class={cn(
 	"border-input bg-background ring-offset-background data-[placeholder]:text-muted-foreground focus:ring-ring flex h-10 w-full items-center justify-between rounded-md border px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 [&>span]:line-clamp-1",
 	className