+++ packages/plugins/io-center/src/ui/components/canvas/canvas-store.svelte.ts
@@ -1,12 +1,12 @@
 import type { ConnectionPort, LogicalConditioner, NodeElement, NodeElementType } from "./types.canvas"
 import { store } from "../../../store.svelte"
 import type { ObjectNodeDataObject } from "../../../ied/object-tree.type"
-import type { LpElement } from "../lp-list/types.lp-list"
+import type { LpElement } from "../right-bar/lp-list/types.lp-list"
 import { NODE_ELEMENT_TYPE, NODE_TYPE } from "@/headless/constants"
 
 class Store {
-	dataObjects = $derived<NodeElement[]>(store.selectedDataObjects.map(dataObjectToNodeElement))
-	logicalConditioners = $derived(store.logicalConditioners.map(LCToNodeElement))
+	dataObjects = $derived<NodeElement[]>(store.selectedDataObject ? [dataObjectToNodeElement(store.selectedDataObject)] : [])
+	logicalConditioners = $derived(store.selectedLogicalConditioners.map(LCToNodeElement))
 	logicalPhysicals = $derived(store.selectedLogicalPhysicals.map(LPToNodeElement))
 	container = $state<HTMLDivElement | null>(null)
 	mousePosition = $state({ x: 0, y: 0 })
@@ -24,7 +24,8 @@ export function LPToNodeElement(lp: LpElement): NodeElement {
 	return {
 		id: lp.id,
 		type: NODE_ELEMENT_TYPE.LP,
-		name: `${lp.type}-${lp.instance}`,
+		name: `${lp.name || lp.type}-${lp.instance}`,
+		title: `${lp.type}-${lp.instance}`,
 		isLinked: lp.isLinked,
 		lnClass: lp.type,
 		numberOfDynamicPorts: lp.numberOfLPDOPorts
@@ -50,8 +51,8 @@ export function dataObjectToNodeElement(dataObject: ObjectNodeDataObject): NodeE
 		dataObject.objectPath.ln?.inst,
 		dataObject.name
 	]
-		.filter(Boolean)
-		.join("-")
+	.filter(Boolean)
+	.join("-")
 
 	const nodeElement = {
 		name: dataObject.name,