+++ packages/plugins/type-designer/src/ui/components/element-card/card-menu.svelte
@@ -32,10 +32,15 @@ const level = $derived(ref ? 'ref' : 'type')
 
 //====== FUNCTIONS ======//
 
+//====== TYPES
+
 function duplicateTypeHandler() {
-	typeElementsStore.duplicateType({
+	typeElementsStore.duplicateElement({
+		element:
+			typeElementsStore.typeElementsPerFamily[type.family][type.id]
+				.element,
 		family: type.family,
-		id: type.id
+		kind: 'type'
 	})
 }
 
@@ -47,6 +52,20 @@ function deleteTypeHandler() {
 
 	sidebarStore.resetCurrentElementType()
 	sidebar.setOpen(false)
+}
+
+//====== REFS
+
+function duplicateRefHandler() {
+	if (!ref) throw new Error('Ref is not defined')
+	typeElementsStore.duplicateElement({
+		element:
+			typeElementsStore.typeElementsPerFamily[type.family][type.id].refs[
+				ref.family
+			][ref.id].element,
+		family: ref.family,
+		kind: 'ref'
+	})
 }
 
 function deleteRefHandler() {
@@ -66,6 +85,7 @@ function deleteRefHandler() {
 	]} />
 {:else if level === 'ref'}
 	<DropdownMenuWorkaround size="sm" actions={[
+		{ label: 'Duplicate', disabled: false, callback: duplicateRefHandler },
 		{ label: 'Delete', disabled: false, callback: deleteRefHandler }
 	]} />
 {/if}